Guard pow() NaN in TAS calculation and fix g_fCoeffP torn reads
Two fixes:

1. AHRS.cpp TAS density-altitude calculation: pow() with a negative or
   zero base and fractional exponent returns NaN (IEEE 754). Bad OAT
   sensor data could make fOAT_k <= 0, and extreme density altitude
   could make the IAS divisor <= 0. Guard both bases so fTAS falls
   through to the simple approximation when inputs are out of range.

2. Globals.h: g_fCoeffP was volatile double — on the 32-bit ESP32
   (dual-core Xtensa), 64-bit reads are not atomic and the value can
   tear between the two 32-bit halves. All writers assign float values
   and the sole reader (DataServer) was casting to float anyway.
   Changed to volatile float for atomic 32-bit access.

+++ software/OnSpeed-Gen3-ESP32/DataServer.cpp
@@ -286,7 +286,7 @@ size_t UpdateLiveDataJson(char * pOut, size_t uOutSize)
 
     const float fPAltFt = SafeJsonFloat(m2ft(g_AHRS.KalmanAlt), 0.0f);
     const float fLatG   = SafeJsonFloat(g_AHRS.AccelLatCorr, 0.0f);
-    const float fCoeffP = SafeJsonFloat((float)g_fCoeffP, 0.0f);
+    const float fCoeffP = SafeJsonFloat(g_fCoeffP, 0.0f);
     const float fPitchRate = SafeJsonFloat(g_AHRS.gPitch, 0.0f);
     const float fDecelRate = SafeJsonFloat(g_Sensors.fDecelRate, 0.0f);
